Move auth.service.UserRegistrationService* -> user.service.UserRegistrationService*
