+++ src/main/java/site/hirecruit/hr/domain/user/service/UserRegistrationServiceImpl.kt
@@ -1,4 +1,4 @@
-package site.hirecruit.hr.domain.auth.service.impl
+package site.hirecruit.hr.domain.user.service
 
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Service
@@ -8,7 +8,6 @@ import site.hirecruit.hr.domain.user.dto.UserRegistrationDto
 import site.hirecruit.hr.domain.user.entity.Role
 import site.hirecruit.hr.domain.user.entity.UserEntity
 import site.hirecruit.hr.domain.auth.repository.UserRepository
-import site.hirecruit.hr.domain.auth.service.UserRegistrationService
 import site.hirecruit.hr.global.event.UserRegistrationEvent
 
 /**
